feat(core): wip: initial attribute aliasing support
