+++ packages/common/private-api.ts
@@ -2,3 +2,4 @@ export * from './src/metadata/metadata-manager';
 export * from './src/metadata/metadata-storage';
 export * from './src/logger/debug-logger';
 export * from './src/error';
+export * from './src/index-options';
